requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
pdf2image
PyMuPDF
crewai
crewai_tools
langchain_huggingface
